@@ -14,58 +14,67 @@ description: 4chan viewer written in Flutter
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
-  flutter: ">=1.10.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.10.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flutter_widgets: ^0.1.12
-  flutter_bloc: ^6.1.1
-  get_it: ^5.0.3
-  http: ^0.12.2
-  dio: ^3.0.10
-  logger_flutter: ^0.7.1
+  flutter_bloc: ^7.0.0
+  get_it: ^6.0.0
+  http: ^0.13.0
+  dio: ^4.0.0-prev2
+  logger: ^1.0.0
+#  logger_flutter: ^0.7.1
 
-  moor: ^3.4.0
-  sqlite3_flutter_libs:
-  sqflite: ^1.3.2+1
+  moor: ^4.1.0
+  sqlite3_flutter_libs: ^0.4.1
+  sqflite: ^2.0.0+3
 
-  worker_manager: ^3.2.1
-  local_auth: ^0.6.3+4
+#  worker_manager: ^3.3.0
+  local_auth: ^1.1.0
 
   path_provider:
   path:
-  filesize: ^1.0.4
-  kt_dart: ^0.8.0
-  equatable: ^1.2.5
-  cupertino_icons: ^1.0.0
+  kt_dart: ^0.9.0
+  equatable: ^2.0.0
+  cupertino_icons: ^1.0.2
 
-  shared_preferences: ^0.5.12+4
-  permission_handler: ^5.0.1+1
-  flutter_downloader: ^1.5.2
-  date_format: ^1.0.9
+  shared_preferences: ^2.0.5
+  permission_handler: ^6.1.1
+  flutter_downloader: ^1.6.0-nullsafety.0
+  date_format: ^2.0.2
 
-  floating_action_bubble: ^1.1.2
-  cached_network_image: ^2.5.0
-  draggable_scrollbar: ^0.0.4
-  pull_to_refresh: ^1.6.3
-  scrollable_positioned_list: ^0.1.8
-  popup_menu: ^1.0.5
-  photo_view: ^0.10.3
-  flutter_html: ^1.1.1
-  html_unescape: ^1.0.2
-  chewie: ^0.12.0
-  video_player: ^1.0.1
-  video_player_platform_interface: ^2.2.0
-  video_thumbnail: ^0.2.5+1
-  carousel_slider: ^2.3.1
-  sliding_sheet: ^0.4.1
+  cached_network_image: ^3.0.0-nullsafety
+  floating_action_bubble:
+    git:
+      url: git@github.com:hovadur/floating_action_bubble.git
+      ref: master
+  draggable_scrollbar:
+    git:
+      url: git@github.com:fzyzcjy/flutter-draggable-scrollbar.git
+      ref: master
+  pull_to_refresh:
+    git:
+      url: git@github.com:fzyzcjy/flutter_pulltorefresh.git
+      ref: master
+
+  scrollable_positioned_list: ^0.2.0-nullsafety.0
+#  popup_menu: ^1.0.5
+  photo_view: ^0.11.1
+  flutter_html: ^2.0.0-nullsafety.0
+  html_unescape: ^2.0.0-nullsafety.0
+  chewie: ^1.0.0
+  video_player: ^2.0.2
+  video_player_platform_interface: ^4.0.0
+  video_thumbnail: ^0.3.1
+  carousel_slider: ^4.0.0-nullsafety.0
+  sliding_sheet: ^0.5.0
 
 dev_dependencies:
-  moor_generator:
-  build_runner:
+  moor_generator: ^4.2.0
+  build_runner: ^1.12.2
 #dev_dependencies:
 #  flutter_test:
 #    sdk: flutter
