@@ -1,27 +1,26 @@
 import 'package:equatable/equatable.dart';
-import 'package:flutter/foundation.dart';
 
 abstract class ChanState extends Equatable {
   final bool showSearchBar;
 
   const ChanState({
-    @required this.showSearchBar,
+    this.showSearchBar = false,
   });
 
   @override
-  List<Object> get props => [showSearchBar];
+  List<Object?> get props => [showSearchBar];
 }
 
 abstract class ChanStateContent extends ChanState {
   final bool showLazyLoading;
 
   const ChanStateContent({
-    @required showSearchBar,
-    @required this.showLazyLoading,
+    required showSearchBar,
+    required this.showLazyLoading,
   }) : super(showSearchBar: showSearchBar);
 
   @override
-  List<Object> get props => super.props..addAll([showLazyLoading]);
+  List<Object?> get props => super.props..addAll([showLazyLoading]);
 }
 
 class ChanStateLoading extends ChanState {}
@@ -32,7 +31,7 @@ class ChanStateError extends ChanState {
   ChanStateError(this.message);
 
   @override
-  List<Object> get props => [message];
+  List<Object?> get props => [message];
 }
 
 class ChanStateNotAuthorized extends ChanState {}
