@@ -1,12 +1,12 @@
 import 'package:flutter/foundation.dart';
 
 class AppException implements Exception {
-  final Exception exception;
+  final Exception? exception;
   final String message;
 
   const AppException({
-    @required this.exception,
-    @required this.message,
+    required this.exception,
+    required this.message,
   });
 }
 
@@ -15,7 +15,7 @@ class HttpException implements Exception {
   final int errorCode;
 
   const HttpException({
-    @required this.message,
-    @required this.errorCode,
+    required this.message,
+    required this.errorCode,
   });
 }
