@@ -1,10 +1,9 @@
 import 'package:equatable/equatable.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter_chan_viewer/data/local/moor_db.dart';
+import 'package:flutter_chan_viewer/models/helper/chan_post_base.dart';
 import 'package:flutter_chan_viewer/models/ui/thread_item.dart';
 import 'package:flutter_chan_viewer/repositories/cache_directive.dart';
 import 'package:flutter_chan_viewer/utils/chan_util.dart';
-import 'package:flutter_chan_viewer/models/helper/chan_post_base.dart';
 import 'package:path/path.dart';
 
 class PostItem extends ChanPostBase with EquatableMixin {
@@ -12,22 +11,22 @@ class PostItem extends ChanPostBase with EquatableMixin {
   final List<int> repliesTo;
   final List<PostItem> repliesFrom;
   final bool isHidden;
-  ThreadItem thread;
+  ThreadItem? thread;
 
   bool get hasReplies => repliesFrom.isNotEmpty;
 
   PostItem({
-    @required boardId,
-    @required threadId,
-    @required timestamp,
-    @required subtitle,
-    @required htmlContent,
-    @required filename,
-    @required imageId,
-    @required extension,
-    @required this.postId,
-    @required this.repliesTo,
-    @required this.repliesFrom,
+    required boardId,
+    required threadId,
+    required timestamp,
+    required subtitle,
+    required htmlContent,
+    required filename,
+    required imageId,
+    required extension,
+    required this.postId,
+    required this.repliesTo,
+    required this.repliesFrom,
     this.isHidden = false,
     this.thread,
   }) : super(
@@ -93,7 +92,7 @@ class PostItem extends ChanPostBase with EquatableMixin {
         isHidden: this.isHidden,
       );
 
-  factory PostItem.fromTableData(PostsTableData entry, {ThreadItem thread}) => PostItem(
+  factory PostItem.fromTableData(PostsTableData entry, {ThreadItem? thread}) => PostItem(
         boardId: entry.boardId,
         threadId: entry.threadId,
         postId: entry.postId,
@@ -104,25 +103,25 @@ class PostItem extends ChanPostBase with EquatableMixin {
         imageId: entry.imageId,
         extension: entry.extension,
         repliesTo: ChanUtil.getPostReferences(entry.content),
-        isHidden: entry.isHidden,
+        isHidden: entry.isHidden ?? false,
         repliesFrom: [],
         thread: thread,
       );
 
   PostItem copyWith({
-    int postId,
-    List<int> repliesTo,
-    List<PostItem> repliesFrom,
-    bool isHidden,
-    ThreadItem thread,
-    String boardId,
-    int threadId,
-    int timestamp,
-    String subtitle,
-    String htmlContent,
-    String filename,
-    String imageId,
-    String extension,
+    int? postId,
+    List<int>? repliesTo,
+    List<PostItem>? repliesFrom,
+    bool? isHidden,
+    ThreadItem? thread,
+    String? boardId,
+    int? threadId,
+    int? timestamp,
+    String? subtitle,
+    String? htmlContent,
+    String? filename,
+    String? imageId,
+    String? extension,
   }) {
     return new PostItem(
       postId: postId ?? this.postId,
@@ -142,5 +141,5 @@ class PostItem extends ChanPostBase with EquatableMixin {
   }
 
   @override
-  List<Object> get props => super.props + [postId, repliesTo, repliesFrom, thread, isHidden];
+  List<Object?> get props => super.props + [postId, repliesTo, repliesFrom, thread, isHidden];
 }
