@@ -15,12 +15,12 @@ import 'package:flutter_chan_viewer/utils/extensions.dart';
 import 'package:flutter_chan_viewer/utils/preferences.dart';
 
 class BoardDetailBloc extends BaseBloc<ChanEvent, ChanState> {
-  final ChanRepository _repository = getIt<ChanRepository>();
+  final ChanRepository? _repository = getIt<ChanRepository>();
 
   BoardDetailBloc(this.boardId) : super(ChanStateLoading());
 
-  final String boardId;
-  BoardDetailModel _boardDetailModel;
+  final String? boardId;
+  BoardDetailModel? _boardDetailModel;
   bool _isFavorite = false;
 
   @override
@@ -28,22 +28,22 @@ class BoardDetailBloc extends BaseBloc<ChanEvent, ChanState> {
     try {
       if (event is ChanEventFetchData) {
         yield ChanStateLoading();
-        List<String> favoriteBoards = Preferences.getStringList(Preferences.KEY_FAVORITE_BOARDS);
+        List<String?> favoriteBoards = Preferences.getStringList(Preferences.KEY_FAVORITE_BOARDS);
         _isFavorite = favoriteBoards.contains(boardId);
 
-        _boardDetailModel = await _repository.fetchCachedBoardDetail(boardId);
+        _boardDetailModel = await _repository!.fetchCachedBoardDetail(boardId);
         if (_boardDetailModel != null) {
           yield _buildContentState(true);
         }
 
-        _boardDetailModel = await _repository.fetchRemoteBoardDetail(boardId);
+        _boardDetailModel = await _repository!.fetchRemoteBoardDetail(boardId);
         yield _buildContentState(false);
       } else if (event is ChanEventSearch) {
         searchQuery = event.query;
         add(ChanEventFetchData());
       } else if (event is BoardDetailEventToggleFavorite) {
         _isFavorite = !_isFavorite;
-        List<String> favoriteBoards = Preferences.getStringList(Preferences.KEY_FAVORITE_BOARDS);
+        List<String?> favoriteBoards = Preferences.getStringList(Preferences.KEY_FAVORITE_BOARDS);
         favoriteBoards.removeWhere((value) => value == boardId);
         if (_isFavorite) {
           favoriteBoards.add(boardId);
@@ -63,11 +63,11 @@ class BoardDetailBloc extends BaseBloc<ChanEvent, ChanState> {
   BoardDetailStateContent _buildContentState(bool showLazyLoading) {
     List<ThreadItem> threads;
     if (searchQuery.isNotNullNorEmpty) {
-      List<ThreadItem> titleMatchThreads = _boardDetailModel.threads.where((thread) => (thread.subtitle ?? "").containsIgnoreCase(searchQuery)).toList();
-      List<ThreadItem> bodyMatchThreads = _boardDetailModel.threads.where((thread) => (thread.content ?? "").containsIgnoreCase(searchQuery)).toList();
+      List<ThreadItem> titleMatchThreads = _boardDetailModel!.threads.where((thread) => (thread.subtitle ?? "").containsIgnoreCase(searchQuery)).toList();
+      List<ThreadItem> bodyMatchThreads = _boardDetailModel!.threads.where((thread) => (thread.content ?? "").containsIgnoreCase(searchQuery)).toList();
       threads = LinkedHashSet<ThreadItem>.from(titleMatchThreads + bodyMatchThreads).toList();
     } else {
-      threads = _boardDetailModel.threads;
+      threads = _boardDetailModel!.threads;
     }
 
     return BoardDetailStateContent(
