@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'dart:collection';
 
-import 'package:bloc/bloc.dart';
 import 'package:flutter_chan_viewer/bloc/chan_event.dart';
 import 'package:flutter_chan_viewer/bloc/chan_state.dart';
 import 'package:flutter_chan_viewer/data/remote/app_exception.dart';
@@ -12,17 +11,17 @@ import 'package:flutter_chan_viewer/pages/favorites/bloc/favorites_event.dart';
 import 'package:flutter_chan_viewer/repositories/chan_repository.dart';
 import 'package:flutter_chan_viewer/utils/chan_logger.dart';
 import 'package:flutter_chan_viewer/utils/chan_util.dart';
+import 'package:flutter_chan_viewer/utils/extensions.dart';
 import 'package:flutter_chan_viewer/utils/preferences.dart';
 import 'package:permission_handler/permission_handler.dart';
-import 'package:flutter_chan_viewer/utils/extensions.dart';
 
 import 'favorites_state.dart';
 
 class FavoritesBloc extends BaseBloc<ChanEvent, ChanState> {
-  final ChanRepository _repository = getIt<ChanRepository>();
+  final ChanRepository? _repository = getIt<ChanRepository>();
   static const int DETAIL_REFRESH_TIMEOUT = 60 * 1000; // 60 seconds
-  List<FavoritesThreadWrapper> _favoriteThreads = List<FavoritesThreadWrapper>();
-  List<FavoritesThreadWrapper> _customThreads = List<FavoritesThreadWrapper>();
+  List<FavoritesThreadWrapper> _favoriteThreads = <FavoritesThreadWrapper>[];
+  List<FavoritesThreadWrapper> _customThreads = <FavoritesThreadWrapper>[];
   int _currentFavoritesRefreshIndex = 0;
   int _lastDetailRefreshTimestamp = 0;
 
@@ -39,15 +38,15 @@ class FavoritesBloc extends BaseBloc<ChanEvent, ChanState> {
           return;
         }
 
-        List<ThreadDetailModel> threads = await _repository.getFavoriteThreads();
+        List<ThreadDetailModel> threads = await _repository!.getFavoriteThreads();
         bool showNsfw = Preferences.getBool(Preferences.KEY_SETTINGS_SHOW_NSFW, def: false);
         if (!showNsfw) {
-          List<String> sfwBoardIds = (await _repository.fetchCachedBoardList(false)).boards.map((board) => board.boardId).toList();
-          threads.removeWhere((model) => !sfwBoardIds.contains(model.thread.boardId));
+          List<String?> sfwBoardIds = (await _repository!.fetchCachedBoardList(false))!.boards.map((board) => board.boardId).toList();
+          threads.removeWhere((model) => !sfwBoardIds.contains(model.thread!.boardId));
         }
         _favoriteThreads = threads.map((e) => FavoritesThreadWrapper(e)).toList();
         _currentFavoritesRefreshIndex = 0;
-        _customThreads = (await _repository.getCustomThreads())
+        _customThreads = (await _repository!.getCustomThreads())
             .map((thread) => FavoritesThreadWrapper(
                   ThreadDetailModel.fromThreadAndPosts(thread, []),
                   isCustom: true,
@@ -64,18 +63,18 @@ class FavoritesBloc extends BaseBloc<ChanEvent, ChanState> {
         }
       } else if (event is FavoritesEventFetchDetail) {
         ThreadDetailModel cachedThread = _favoriteThreads[_currentFavoritesRefreshIndex].threadDetailModel;
-        ThreadDetailModel refreshedThread;
+        ThreadDetailModel? refreshedThread;
 
         _favoriteThreads[_currentFavoritesRefreshIndex] = FavoritesThreadWrapper(cachedThread, isLoading: true);
         yield _buildContentState(showLazyLoading: true);
 
         int newReplies = _favoriteThreads[_currentFavoritesRefreshIndex].newReplies;
         try {
-          refreshedThread = await _repository.fetchRemoteThreadDetail(cachedThread.thread.boardId, cachedThread.thread.threadId, false);
-          int newMedia = refreshedThread.thread.images - cachedThread.thread.images;
-          newReplies += refreshedThread.thread.replies - cachedThread.thread.replies;
+          refreshedThread = await _repository!.fetchRemoteThreadDetail(cachedThread.thread!.boardId, cachedThread.thread!.threadId, false);
+          int newMedia = refreshedThread!.thread!.images! - cachedThread.thread!.images!;
+          newReplies += refreshedThread.thread!.replies! - cachedThread.thread!.replies!;
           if (newMedia > 0) {
-            _repository.downloadAllMedia(refreshedThread);
+            _repository!.downloadAllMedia(refreshedThread);
           }
         } on HttpException catch (e, stackTrace) {
           // ChanLogger.v("Thread not found. Probably offline. Ignoring");
@@ -104,10 +103,10 @@ class FavoritesBloc extends BaseBloc<ChanEvent, ChanState> {
     List<FavoritesThreadWrapper> favoriteThreads;
     if (searchQuery.isNotNullNorEmpty) {
       List<FavoritesThreadWrapper> titleMatchThreads = _favoriteThreads.where((thread) {
-        return (thread.threadDetailModel.thread.subtitle ?? "").containsIgnoreCase(searchQuery);
+        return (thread.threadDetailModel.thread!.subtitle ?? "").containsIgnoreCase(searchQuery);
       }).toList();
       List<FavoritesThreadWrapper> bodyMatchThreads = _favoriteThreads.where((thread) {
-        return (thread.threadDetailModel.thread.content ?? "").containsIgnoreCase(searchQuery);
+        return (thread.threadDetailModel.thread!.content ?? "").containsIgnoreCase(searchQuery);
       }).toList();
       favoriteThreads = LinkedHashSet<FavoritesThreadWrapper>.from(titleMatchThreads + bodyMatchThreads).toList();
     } else {
@@ -121,10 +120,10 @@ class FavoritesBloc extends BaseBloc<ChanEvent, ChanState> {
     List<FavoritesThreadWrapper> customThreads;
     if (searchQuery.isNotNullNorEmpty) {
       List<FavoritesThreadWrapper> titleMatchThreads = _customThreads.where((thread) {
-        return (thread.threadDetailModel.thread.subtitle ?? "").containsIgnoreCase(searchQuery);
+        return (thread.threadDetailModel.thread!.subtitle ?? "").containsIgnoreCase(searchQuery);
       }).toList();
       List<FavoritesThreadWrapper> bodyMatchThreads = _customThreads.where((thread) {
-        return (thread.threadDetailModel.thread.content ?? "").containsIgnoreCase(searchQuery);
+        return (thread.threadDetailModel.thread!.content ?? "").containsIgnoreCase(searchQuery);
       }).toList();
       customThreads = LinkedHashSet<FavoritesThreadWrapper>.from(titleMatchThreads + bodyMatchThreads).toList();
     } else {
