@@ -1,6 +1,6 @@
 class CacheDirective {
-  final String boardId;
-  final int threadId;
+  final String? boardId;
+  final int? threadId;
   static const String DIVIDER = "_";
 
   CacheDirective(this.boardId, this.threadId);
