@@ -1,7 +1,7 @@
 import 'package:logger/logger.dart';
 
 class ChanLogger {
-  static Logger _instance;
+  static late Logger _instance;
   static bool _initialized = false;
 
   static void _initialize() {
@@ -21,32 +21,32 @@ class ChanLogger {
     if (!_initialized) _initialize();
   }
 
-  static void v(dynamic message, [dynamic error, StackTrace stackTrace]) {
+  static void v(dynamic message, [dynamic error, StackTrace? stackTrace]) {
     _checkInitialized();
     _instance.v(message, error, stackTrace);
   }
 
-  static void d(dynamic message, [dynamic error, StackTrace stackTrace]) {
+  static void d(dynamic message, [dynamic error, StackTrace? stackTrace]) {
     _checkInitialized();
     _instance.d(message, error, stackTrace);
   }
 
-  static void i(dynamic message, [dynamic error, StackTrace stackTrace]) {
+  static void i(dynamic message, [dynamic error, StackTrace? stackTrace]) {
     _checkInitialized();
     _instance.i(message, error, stackTrace);
   }
 
-  static void w(dynamic message, [dynamic error, StackTrace stackTrace]) {
+  static void w(dynamic message, [dynamic error, StackTrace? stackTrace]) {
     _checkInitialized();
     _instance.w(message, error, stackTrace);
   }
 
-  static void e(dynamic message, [dynamic error, StackTrace stackTrace]) {
+  static void e(dynamic message, [dynamic error, StackTrace? stackTrace]) {
     _checkInitialized();
     _instance.e(message, error, stackTrace);
   }
 
-  static void wtf(dynamic message, [dynamic error, StackTrace stackTrace]) {
+  static void wtf(dynamic message, [dynamic error, StackTrace? stackTrace]) {
     _checkInitialized();
     _instance.wtf(message, error, stackTrace);
   }
