@@ -10,8 +10,8 @@ class BoardsDao extends DatabaseAccessor<MoorDB> with _$BoardsDaoMixin {
 
 //  Stream<List<PostsTableData>> get allActiveBoardItemsStream => select(boardsTable).watch();
 
-  Future<BoardsTableData> getBoardById(String boardId) {
-    return (select(boardsTable)..where((board) => board.boardId.equals(boardId))).getSingle();
+  Future<BoardsTableData?> getBoardById(String boardId) {
+    return (select(boardsTable)..where((board) => board.boardId.equals(boardId))).getSingleOrNull();
   }
 
   Future<List<BoardsTableData>> getBoardItems(bool includeNsfw) {
