@@ -60,7 +60,7 @@ class MainApp extends StatelessWidget {
           ),
         );
       } else {
-        return BasePageState.buildErrorScreen(context, (state as ChanStateError)?.message);
+        return BasePageState.buildErrorScreen(context, (state as ChanStateError).message);
       }
     });
   }
