@@ -12,17 +12,17 @@ class ThreadsTable extends ChanBaseTable {
 
   IntColumn get threadId => integer()();
 
-  IntColumn get lastModified => integer().withDefault(const Constant(0))();
+  IntColumn get lastModified => integer().nullable().withDefault(const Constant(0))();
 
-  IntColumn get selectedPostId => integer().withDefault(const Constant(-1))();
+  IntColumn get selectedPostId => integer().nullable().withDefault(const Constant(-1))();
 
-  BoolColumn get isFavorite => boolean()();
+  BoolColumn get isFavorite => boolean().nullable()();
 
-  IntColumn get onlineState => intEnum<OnlineState>()();
+  IntColumn get onlineState => intEnum<OnlineState>().nullable()();
 
-  IntColumn get replyCount => integer()();
+  IntColumn get replyCount => integer().nullable()();
 
-  IntColumn get imageCount => integer()();
+  IntColumn get imageCount => integer().nullable()();
 
-  IntColumn get unreadRepliesCount => integer().withDefault(const Constant(0))();
+  IntColumn get unreadRepliesCount => integer().nullable().withDefault(const Constant(0))();
 }
