@@ -31,9 +31,9 @@ class BoardItem extends Equatable {
   factory BoardItem.fromTableData(BoardsTableData entry) => BoardItem(
         entry.boardId,
         entry.title,
-        entry.workSafe ? 1 : 0,
+        entry.workSafe! ? 1 : 0,
       );
 
   @override
-  List<Object> get props => [boardId, title, workSafe];
+  List<Object?> get props => [boardId, title, workSafe];
 }
