@@ -3,7 +3,7 @@ import 'package:flutter_chan_viewer/bloc/chan_event.dart';
 import 'package:flutter_chan_viewer/bloc/chan_state.dart';
 
 abstract class BaseBloc<E extends ChanEvent, S extends ChanState> extends Bloc<E, S> {
-  BaseBloc(ChanState state) : super(state);
+  BaseBloc(ChanState state) : super(state as S);
 
   bool _showSearchBar = false;
   String searchQuery = "";
@@ -11,7 +11,7 @@ abstract class BaseBloc<E extends ChanEvent, S extends ChanState> extends Bloc<E
   get showSearchBar => _showSearchBar;
 
   @override
-  Stream<S> mapEventToState(E event) {
+  Stream<S> mapEventToState(E event) async* {
     if (event is ChanEventSearch) {
       searchQuery = event.query;
     } else if (event is ChanEventShowSearch) {
