@@ -1,10 +1,10 @@
 import 'dart:math';
 
 import 'package:date_format/date_format.dart';
+import 'package:flutter_chan_viewer/utils/extensions.dart';
 import 'package:html/dom.dart';
 import 'package:html/parser.dart';
 import 'package:html_unescape/html_unescape.dart';
-import 'package:flutter_chan_viewer/utils/extensions.dart';
 
 class ChanUtil {
   var style = '';
@@ -12,9 +12,9 @@ class ChanUtil {
   static const int MAX_TEXT_LENGTH = 300;
   static HtmlUnescape unescaper = HtmlUnescape();
 
-  static String unescapeHtml(String raw) => unescaper.convert(raw ?? "");
+  static String unescapeHtml(String? raw) => unescaper.convert(raw ?? "");
 
-  static String getReadableHtml(String htmlContent, bool truncate) {
+  static String? getReadableHtml(String? htmlContent, bool truncate) {
     if (htmlContent == null) {
       htmlContent = 'null';
     } else if (truncate && htmlContent.length > IDEAL_TEXT_LENGTH) {
@@ -24,24 +24,24 @@ class ChanUtil {
     return htmlContent;
   }
 
-  static String getPlainString(String htmlContent) {
+  static String getPlainString(String? htmlContent) {
     String rawContent = "";
     if (htmlContent.isNotNullNorEmpty) {
-      Document document = parse(htmlContent.replaceAll("<br>", " ").replaceAll("</p><p>", " "));
-      rawContent = parse(document.body.text).documentElement.text;
+      Document document = parse(htmlContent!.replaceAll("<br>", " ").replaceAll("</p><p>", " "));
+      rawContent = parse(document.body!.text).documentElement!.text;
     }
     return rawContent;
   }
 
-  static List<int> getPostReferences(String content) {
+  static List<int> getPostReferences(String? content) {
     Document document = parse(content ?? "");
     List<Element> links = document.querySelectorAll('body > a');
 //    List<Map<String, dynamic>> linkMap = [];
     List<int> postIds = [];
 
     for (var link in links) {
-      int replyTo = getPostIdFromUrl(link.attributes['href']);
-      if (replyTo != null) {
+      int replyTo = getPostIdFromUrl(link.attributes['href']!);
+      if (replyTo > -1) {
         postIds.add(replyTo);
       }
 //      linkMap.add({
@@ -53,7 +53,7 @@ class ChanUtil {
     return postIds;
   }
 
-  static int getPostIdFromUrl(String url) => url.startsWith("#p") ? int.parse(url.substring(2)) : null;
+  static int getPostIdFromUrl(String url) => url.startsWith("#p") ? int.parse(url.substring(2)) : -1;
 
   static String getHumanDate(int timestamp) {
     return formatDate(DateTime.fromMillisecondsSinceEpoch(timestamp * 1000), [mm, '-', dd, ' ', HH, ':', nn, ':', ss]);
